+++ CHANGES.md
@@ -8,6 +8,7 @@
 
 ## Bug fixes
 * Fix small bug in global data flow analysis (#1768)
+* Runtime: no longer leak channels
 
 # 5.9.1 (02-12-2024) - Lille